[Breaking] Rename /network endpoints to /chain
